refactor: move payment method tracking to payments plugin
