+++ imports/plugins/core/core/server/Reaction/core.js
@@ -15,7 +15,6 @@ import {
   customPublishedProductVariantFields,
   functionsByType,
   mutations,
-  paymentMethods,
   queries,
   resolvers,
   schemas
@@ -116,15 +115,6 @@ export default {
         }
         functionsByType[type].push(...packageInfo.functionsByType[type]);
       });
-    }
-
-    if (packageInfo.paymentMethods) {
-      for (const paymentMethod of packageInfo.paymentMethods) {
-        paymentMethods[paymentMethod.name] = {
-          ...paymentMethod,
-          pluginName: packageInfo.name
-        };
-      }
     }
 
     if (packageInfo.catalog) {
