+++ imports/plugins/core/payments/server/no-meteor/queries/availablePaymentMethods.js
@@ -1,5 +1,5 @@
 import ReactionError from "@reactioncommerce/reaction-error";
-import { paymentMethods } from "/imports/plugins/core/core/server/no-meteor/pluginRegistration";
+import { paymentMethods } from "../registration";
 
 /**
  * @name availablePaymentMethods
